requirements.txt
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.23.1
azure-data-tables==12.3.0
azure-functions==1.11.0
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
isodate==0.6.1
msrest==0.6.21
oauthlib==3.2.0
Pillow==9.1.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
typing-extensions==4.2.0
urllib3==1.26.9
